medical-rag-ai/requirements.txt
fastapi
uvicorn
streamlit
python-multipart
pydantic>=2.0
langchain
langchain-community
langchain-huggingface
langchain-openai
sentence-transformers
faiss-cpu
torch
transformers
pillow
pypdf
python-dotenv
httpx
tiktoken
langchain-google-genai
google-generativeai

